refactor(aggregate): move aggregate-only typing surface out of kernel
