+++ packages/aggregate/src/index.ts
@@ -3,3 +3,11 @@ export * from './createEntity';
 export * from './createMixin';
 export * from './bindContext';
 export * from './naming';
+export type {
+  ResolveEventName,
+  EventEmitterFactory,
+  PackedCommand,
+  PackedCommandWithMeta,
+  ShorthandCommandWithMeta,
+  MapCommandsToPayloads
+} from './types/aggregateTyping';
